+++ Mile.Helpers/Mile.Helpers.h
@@ -80,6 +80,201 @@ typedef enum MILE_PREFERRED_APP_MODE
 } MILE_PREFERRED_APP_MODE;
 
 /**
+ * @brief Values that specify the type of privileges being assigned to or
+ *        retrieved from an access token.
+ * @remark Definitions are synchronized from in wdm.h of Windows Driver Kit.
+ *         These must be converted to LUIDs before use.
+*/
+typedef enum MILE_TOKEN_PRIVILEGE_LUID_CLASS
+{
+    /**
+     * @brief The minimal value for this enumeration.
+    */
+    MILE_TOKEN_MIN_WELL_KNOWN_PRIVILEGE = 2,
+
+    /**
+     * @brief Create a token object.
+    */
+    MILE_TOKEN_CREATE_TOKEN_PRIVILEGE = 2,
+
+    /**
+     * @brief Replace a process-level token.
+    */
+    MILE_TOKEN_ASSIGNPRIMARYTOKEN_PRIVILEGE = 3,
+
+    /**
+     * @brief Lock pages in memory.
+    */
+    MILE_TOKEN_LOCK_MEMORY_PRIVILEGE = 4,
+
+    /**
+     * @brief Adjust memory quotas for a process.
+    */
+    MILE_TOKEN_INCREASE_QUOTA_PRIVILEGE = 5,
+
+    /**
+     * @brief Add workstations to domain.
+    */
+    MILE_TOKEN_MACHINE_ACCOUNT_PRIVILEGE = 6,
+
+    /**
+     * @brief Act as part of the operating system.
+    */
+    MILE_TOKEN_TCB_PRIVILEGE = 7,
+
+    /**
+     * @brief Manage auditing and security log.
+    */
+    MILE_TOKEN_SECURITY_PRIVILEGE = 8,
+
+    /**
+     * @brief Take ownership of files or other objects.
+    */
+    MILE_TOKEN_TAKE_OWNERSHIP_PRIVILEGE = 9,
+
+    /**
+     * @brief Load and unload device drivers.
+    */
+    MILE_TOKEN_LOAD_DRIVER_PRIVILEGE = 10,
+
+    /**
+     * @brief Profile system performance.
+    */
+    MILE_TOKEN_SYSTEM_PROFILE_PRIVILEGE = 11,
+
+    /**
+     * @brief Change the system time.
+    */
+    MILE_TOKEN_SYSTEMTIME_PRIVILEGE = 12,
+
+    /**
+     * @brief Profile single process.
+    */
+    MILE_TOKEN_PROF_SINGLE_PROCESS_PRIVILEGE = 13,
+
+    /**
+     * @brief Increase scheduling priority.
+    */
+    MILE_TOKEN_INC_BASE_PRIORITY_PRIVILEGE = 14,
+
+    /**
+     * @brief Create a pagefile.
+    */
+    MILE_TOKEN_CREATE_PAGEFILE_PRIVILEGE = 15,
+
+    /**
+     * @brief Create permanent shared objects.
+    */
+    MILE_TOKEN_CREATE_PERMANENT_PRIVILEGE = 16,
+
+    /**
+     * @brief Back up files and directories.
+    */
+    MILE_TOKEN_BACKUP_PRIVILEGE = 17,
+
+    /**
+     * @brief Restore files and directories.
+    */
+    MILE_TOKEN_RESTORE_PRIVILEGE = 18,
+
+    /**
+     * @brief Shut down the system.
+    */
+    MILE_TOKEN_SHUTDOWN_PRIVILEGE = 19,
+
+    /**
+     * @brief Debug programs.
+    */
+    MILE_TOKEN_DEBUG_PRIVILEGE = 20,
+
+    /**
+     * @brief Generate security audits.
+    */
+    MILE_TOKEN_AUDIT_PRIVILEGE = 21,
+
+    /**
+     * @brief Modify firmware environment values.
+    */
+    MILE_TOKEN_SYSTEM_ENVIRONMENT_PRIVILEGE = 22,
+
+    /**
+     * @brief Bypass traverse checking.
+    */
+    MILE_TOKEN_CHANGE_NOTIFY_PRIVILEGE = 23,
+
+    /**
+     * @brief Force shutdown from a remote system.
+    */
+    MILE_TOKEN_REMOTE_SHUTDOWN_PRIVILEGE = 24,
+
+    /**
+     * @brief Remove computer from docking station.
+    */
+    MILE_TOKEN_UNDOCK_PRIVILEGE = 25,
+
+    /**
+     * @brief Synchronize directory service data.
+    */
+    MILE_TOKEN_SYNC_AGENT_PRIVILEGE = 26,
+
+    /**
+     * @brief Enable computer and user accounts to be trusted for delegation.
+    */
+    MILE_TOKEN_ENABLE_DELEGATION_PRIVILEGE = 27,
+
+    /**
+     * @brief Manage the files on a volume.
+    */
+    MILE_TOKEN_MANAGE_VOLUME_PRIVILEGE = 28,
+
+    /**
+     * @brief Impersonate a client after authentication.
+    */
+    MILE_TOKEN_IMPERSONATE_PRIVILEGE = 29,
+
+    /**
+     * @brief Create global objects.
+    */
+    MILE_TOKEN_CREATE_GLOBAL_PRIVILEGE = 30,
+
+    /**
+     * @brief Access Credential Manager as a trusted caller.
+    */
+    MILE_TOKEN_TRUSTED_CREDMAN_ACCESS_PRIVILEGE = 31,
+
+    /**
+     * @brief Modify an object label.
+    */
+    MILE_TOKEN_RELABEL_PRIVILEGE = 32,
+
+    /**
+     * @brief Increase a process working set.
+    */
+    MILE_TOKEN_INC_WORKING_SET_PRIVILEGE = 33,
+
+    /**
+     * @brief Change the time zone.
+    */
+    MILE_TOKEN_TIME_ZONE_PRIVILEGE = 34,
+
+    /**
+     * @brief Create symbolic links.
+    */
+    MILE_TOKEN_CREATE_SYMBOLIC_LINK_PRIVILEGE = 35,
+
+    /**
+     * @brief Impersonate other users.
+    */
+    MILE_TOKEN_DELEGATE_SESSION_USER_IMPERSONATE_PRIVILEGE = 36,
+
+    /**
+     * @brief The maximum value for this enumeration.
+    */
+    MILE_TOKEN_MAX_WELL_KNOWN_PRIVILEGE = 36
+
+} MILE_TOKEN_PRIVILEGE_LUID_CLASS;
+
+/**
  * @brief Indicates if the current OS version matches, or is greater than, the
  *        provided version information.
  * @param Major The major version number of the operating system.