+++ ReleaseNotes.md
@@ -16,3 +16,4 @@
 - Add MileAllowDarkModeForApp function.
 - Add MileGetDefaultBackgroundColorValue function.
 - Add MileEnableImmersiveDarkModeForWindow function.
+- Add MILE_TOKEN_PRIVILEGE_LUID_CLASS enum.