+++ ReleaseNotes.md
@@ -1,8 +1,7 @@
 ﻿# Mile.Windows.Helpers Release Notes
 
-**Mile.Windows.Helpers 1.0 (1.0.11.0)**
+**Mile.Windows.Helpers 1.0 (1.0.12.0)**
 
-- Provide static library mode.
 - Provide NuGet package.
 - Add MileIsWindowsVersionAtLeast function.
 - Add MileEnablePerMonitorDialogScaling function.